Moves computeResources to a separate function
The computeResources function should not have to make any calls
to the Kubernetes API to check state - that is, it should be able
to determine all required resources from the Cluster object.
So, we force this by moving it to a function, as opposed to a
method, so it does not have access to the k8s clientset.