Removes cluster ID prefix from Kubernetes resources
It is entirely possible for the ID generator to give us an ID
which starts with a number, which is prohibited for names of
Kubernetes resources. This changeset removes the prefix from
resources - it is unnecessary, as we have namespaces and
labelling, which prevents the situation where we are creating
Kubernetes resources whose name starts with an integer.