service: split kvmconfig and pod frameworks files
Goal here is to make it easier to copy things when introducing new
version bundle.

CRD framework was renamed to KVMConfig framework because it watches
kvmconfigs, not customresourcedefinitions. Otherwise it's copy/pasting
to separate file.

Next steps is to route initCtx together with resources (requires changes
in operatorkit), and introduce unversioned packages (like
keyunversioned).