Allocate a unique index for each node and store it in status
It is occasionally necessary to have a guaranteed-to-be-unique index for each
node that is as small integer as possible. This can be used to
deterministically sort nodes but it's also useful for other purposes.

One use case now is ability to provide deterministic iSCSI initiator names for
each node in a way that those names can be pre-configured in storage system but
so that our cluster worker node set can still be dynamic - especially during
upgrades.

This is achieved by allocating each node an index that is stored in provider
specific status section. This index is as small as possible at the time. When a
node is removed from a cluster, there is a "hole" in the node indexes, but next
new node will then get this index.